requirements.txt
# flask
strawberry-graphql
psycopg[binary]
fastapi[standard]